Refactor large test suites into modular files
